+++ Backend/GranReserva/Domain/Entities/RewardExchange.cs
@@ -4,7 +4,7 @@ namespace Domain.Entities
     {
         public int Id { get; set; }
         public DateTime Date { get; set; } = DateTime.UtcNow;
-        public int PointsUsed { get; set; }
+        public long PointsUsed { get; set; }
         public int ClientId { get; set; }
         public Client? Client { get; set; }
         public int RewardId { get; set; }
